[M1386-2]

+++ src/api/models/mermaid.py
@@ -499,8 +499,8 @@ class SampleUnit(BaseModel):
 
 
 class Transect(SampleUnit):
-    len_surveyed = models.PositiveSmallIntegerField(
-        verbose_name=_(u'transect length surveyed (m)'))
+    len_surveyed = models.DecimalField(max_digits=4, decimal_places=1,
+                                       verbose_name=_(u'transect length surveyed (m)'))
     reef_slope = models.ForeignKey(
         ReefSlope, on_delete=models.SET_NULL, null=True, blank=True)
