+++ src/api/models/view_models/beltfish.py
@@ -79,9 +79,7 @@ CREATE OR REPLACE VIEW public.vw_beltfish_obs
 
     sample_unit_id = models.UUIDField()
     transect_number = models.PositiveSmallIntegerField()
-    transect_len_surveyed = models.PositiveSmallIntegerField(
-        verbose_name=_("transect length surveyed (m)")
-    )
+    transect_len_surveyed = models.DecimalField(max_digits=4, decimal_places=1, verbose_name=_("transect length surveyed (m)"))
     transect_width_name = models.CharField(max_length=100, null=True, blank=True)
     reef_slope = models.CharField(max_length=50)
     fish_family = models.CharField(max_length=100, null=True, blank=True)
@@ -242,9 +240,7 @@ ON (beltfish_su.pseudosu_id = beltfish_obs.pseudosu_id)
     sample_unit_ids = JSONField()
     total_abundance = models.PositiveIntegerField()
     transect_number = models.PositiveSmallIntegerField()
-    transect_len_surveyed = models.PositiveSmallIntegerField(
-        verbose_name=_("transect length surveyed (m)")
-    )
+    transect_len_surveyed = models.DecimalField(max_digits=4, decimal_places=1, verbose_name=_("transect length surveyed (m)"))
     transect_width_name = models.CharField(max_length=100, null=True, blank=True)
     reef_slope = models.CharField(max_length=50)
     size_bin = models.CharField(max_length=100)
